server separated from main; first github/workflow setup init
